requirements.txt
discord-py>=2.6.4
python-dotenv>=1.2.1
openai>=1.109.1
pymongo>=4.6.0
pymongo>=4.15.4